requirements.txt
colorama
discord.py
git+https://github.com/skullbite/disputils
aiohttp
requests
pymongo
ast

 